requirements.txt
Django==2.2.1
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
drf-yasg==1.15.0                        # API文档
django-cors-headers==2.4.0              # 跨域 2.4.0
pillow==6.0.0                           # Django ImageField支持库
django-easy-select2==1.5.5              # 多对多字段admin插件
djangorestframework-recursive==0.1.2
django-debug-toolbar==1.11              # debug工具栏
django-jet==1.0.8                       # 自定义admin界面
FeedParser==5.2.1
cryptography==2.6.1                     # 加密
xlrd==1.2.0                             # excel操作
xlwt
Celery==4.3.0                           # 任务队列
django-celery-results==1.1.2
django-celery-beat==1.5.0
sentry-sdk==0.9.0
flower==0.9.3
django-redis-cache
requests
mysqlclient
markdown
django-rest-auth
django-redis
drf-extensions
